requirements_test.txt
# test 不限制版本，不同python版本就安装最高版本
pytest
pytest-env
pytest-cov
pytest-django
